Extracted IPFS Gateway logic to utility function - changed IPFSImage component to handle any img data - added IPFS img handling to OGCards - changed baseUrl for OG card to come form vercel url
